google_cloud/requirements.txt
# =============================================================================
# google_cloud/requirements.txt
# Google Cloud & supporting dependencies for the sensor anomaly system
#
# Install with:
#   pip install -r requirements.txt
# =============================================================================

# ---------------------------------------------------------------------------
# Google Cloud client libraries
# ---------------------------------------------------------------------------
google-cloud-pubsub==2.21.4
google-cloud-bigquery==3.20.1
google-cloud-bigquery-storage==2.25.0
google-cloud-storage==2.16.0
google-cloud-aiplatform==1.54.0
google-cloud-monitoring==2.21.0
google-api-core==2.19.0

# ---------------------------------------------------------------------------
# Apache Beam (Dataflow runner)
# ---------------------------------------------------------------------------
apache-beam[gcp]==2.56.0

# ---------------------------------------------------------------------------
# ML / data processing
# ---------------------------------------------------------------------------
scikit-learn==1.3.2
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
pyarrow==15.0.2

# ---------------------------------------------------------------------------
# FastAPI (Cloud Run service)
# ---------------------------------------------------------------------------
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1

# ---------------------------------------------------------------------------
# Utilities
# ---------------------------------------------------------------------------
python-dotenv==1.0.1
requests==2.32.2
